TheOnePiece
description;
One Piece is a Japanese manga series written and illustrated by Eiichiro Oda. now we are creating a meme token that is one piece which is present in the solana networked token.

A memecoin is a joke that turns into a crypto coin, most of the time with a high market capitalization. The One Piece used a meme to get some awareness of the Solana community.