Add bitadx
Bitadx is the India’s first cryptocurrency exchange on Solana Blockchain. It is the right platform for trading, no matter if you are an experienced trader or just started. We offer the fastest order execution and competitive rates. With face detection login your experience gets even better! We also offer DEX or LP for new tokens.